Create Backend_ClientNode
Backend which receives the data from the frontend in JSON format. 
Establishes connection with the SQS in order to send the user data.
Clear display of any errors faced, to troubleshoot more easily.
Utilizes flask for connection with the frontend, Cors to setup the server, boto3 for aws python SDK, and JSON for the receipt of data from frontend.