Create main.py
Responsible for handling all interactions with aws components (the SQS & S3) (Sending to and Receiving from AWS services)
Receives the Url and args from the SQS, proceeding to forward that data in JSON format to the crawler for crawling and scraping, and triggers the crawler to work upon receipt of data. 
Uses output from crawler file (crawling_spider.py) to send it to the indexer via another SQS (result_queue). And stores the raw html data on the S3.

Implements two methods of execution, the method intended for production utilizes an infinite loop to continuously poll the SQS for any input from the client (that input is passed to the master then to the crawler*). 
System runs continuously, but polling is done in intervals (periodic).